# Conflicts:
#	src/ImageSharp/Formats/Bmp/BmpConstants.cs
#	src/ImageSharp/Formats/Bmp/BmpDecoder.cs
#	src/ImageSharp/Formats/Bmp/BmpFormat.cs

+++ src/ImageSharp/Processing/Processors/Transforms/Resize/ResizeKernelMap.cs
@@ -1,5 +1,5 @@
 // Copyright (c) Six Labors.
-// Licensed under the Apache License, Version 2.0.
+// Licensed under the Six Labors Split License.
 
 using System;
 using System.Buffers;
@@ -101,7 +101,7 @@ namespace SixLabors.ImageSharp.Processing.Processors.Transforms
         /// Returns a <see cref="ResizeKernel"/> for an index value between 0 and DestinationSize - 1.
         /// </summary>
         [MethodImpl(InliningOptions.ShortMethod)]
-        internal ref ResizeKernel GetKernel(int destIdx) => ref this.kernels[destIdx];
+        internal ref ResizeKernel GetKernel(nint destIdx) => ref this.kernels[destIdx];
 
         /// <summary>
         /// Computes the weights to apply at each pixel when resizing.
